gh-119853: Add Include/refcount.h file
